Move alias settings to sub-tabs
