+++ SeriesMgr/Aliases.py
@@ -1,9 +1,9 @@
 import wx
-import os
-import sys
-import SeriesModel
-import Utils
 from AliasGrid import AliasGrid
+from AliasesName		import AliasesName
+from AliasesLicense		import AliasesLicense
+from AliasesMachine		import AliasesMachine
+from AliasesTeam		import AliasesTeam
 
 def normalizeText( text ):
 	return ', '.join( [t.strip() for t in text.split(',')][:2] )
@@ -11,111 +11,60 @@ def normalizeText( text ):
 class Aliases(wx.Panel):
 	def __init__(self, parent):
 		wx.Panel.__init__(self, parent)
+		self.notebook = wx.Notebook( self )
+		self.notebook.Bind( wx.EVT_NOTEBOOK_PAGE_CHANGED, self.onPageChanging )
 		
-		text =	(
-			'Name Aliases match different name spellings to the same participant.\n'
-			'This can be more convenient than editing race results when the same participant has resullts under different names.\n'
-			'\n'
-			'To create a name Alias, first press the "Add Reference Name" button.\n'
-			'The first column is the name that will appear in Results.'
-			'Then, add Aliases in the next column separated by ";" (semicolons). These are the alternate spellings of the name.\n'
-			'SeriesMgr will match all Aliases to the Reference Name in the Results.\n'
-			'\n'
-			'For example, Reference Name="Bell, Robert", Aliases="Bell, Bobby; Bell, Bob".  Results for the alternate spellings will appear as "Bell, Robert".\n'
-			'Accents and upper/lower case are ignored.\n'
-			'\n'
-			'You can Copy-and-Paste names from the Results without retyping them.  Right-click and Copy the name in the Results page,'
-			'then Paste the name into the Reference Name or Alias field.\n'
-			'Aliases will not be applied until you press the "Refresh" button on the Results screen (or reload).\n'
-			'This allows you to configure many Aliases without having to wait for the Results update after each change.\n'
-		)
-		
-		self.explain = wx.StaticText( self, label=text )
-		self.explain.SetFont( wx.Font((0,15), wx.FONTFAMILY_SWISS, wx.FONTSTYLE_NORMAL, wx.FONTWEIGHT_NORMAL, False) )
-		
-		self.addButton = wx.Button( self, label='Add Reference Name' )
-		self.addButton.Bind( wx.EVT_BUTTON, self.onAddButton )
-		
-		headerNames = ('Name (Last, First)','Aliases separated by ";"')
-		self.itemCur = None
-		self.grid = AliasGrid( self )
-		self.grid.CreateGrid( 0, len(headerNames) )
-		self.grid.SetRowLabelSize( 64 )
-		for col in range(self.grid.GetNumberCols()):
-			self.grid.SetColLabelValue( col, headerNames[col] )
-			
-		sizer = wx.BoxSizer(wx.VERTICAL)
-		sizer.Add(self.explain, 0, flag=wx.ALL, border=4 )
-		sizer.Add(self.addButton, 0, flag=wx.ALL, border = 4)
-		sizer.Add(self.grid, 1, flag=wx.EXPAND|wx.ALL, border = 4)
-		self.SetSizer(sizer)
-	
-	def onAddButton( self, event ):
-		defaultText = ''
-		
-		# Initialize the name from the clipboard.
-		if wx.TheClipboard.Open():
-			do = wx.TextDataObject()
-			if wx.TheClipboard.GetData(do):
-				defaultText = do.GetText()
-			wx.TheClipboard.Close()
+		# Add all the pages to the notebook.
+		self.pages = []
 
-		self.grid.AppendRows( 1 )
-		self.grid.SetCellValue( self.grid.GetNumberRows()-1, 0, defaultText )
-		self.grid.AutoSize()
-		self.GetSizer().Layout()
-		self.grid.MakeCellVisible( self.grid.GetNumberRows()-1, 0 )
-	
-	def getName( self, s ):
-		name = [t.strip() for t in s.split(',')[:2]]
-		if not name or not any(name):
-			return None
-		name.extend( [''] * (2 - len(name)) )
-		return tuple( name )
+		def addPage( page, name ):
+			self.notebook.AddPage( page, name )
+			self.pages.append( page )
+			
+		self.attrClassName = [
+			[ 'aliasesName',	AliasesName,		'Name Aliases' ],
+			[ 'licenseAliases',	AliasesLicense,		'License Aliases' ],
+			[ 'machineAliases',	AliasesMachine,		'Machine Aliases' ],
+			[ 'teamAliases',	AliasesTeam,		'Team Aliases' ],
+		]
 		
+		for i, (a, c, n) in enumerate(self.attrClassName):
+			setattr( self, a, c(self.notebook) )
+			addPage( getattr(self, a), n )
+		
+		self.notebook.SetSelection( 0 )
+		
+		sizer = wx.BoxSizer( wx.VERTICAL )
+		sizer.Add( self.notebook, 1, flag=wx.EXPAND )
+		self.SetSizer( sizer )
+		
+		wx.CallAfter( self.Refresh )
+		
+	def refreshCurrentPage( self ):
+		self.setTitle()
+		self.callPageRefresh( self.notebook.GetSelection() )
+
 	def refresh( self ):
-		model = SeriesModel.model
+		self.refreshCurrentPage()
+
+	def callPageRefresh( self, i ):
+		try:
+			self.pages[i].refresh()
+		except (AttributeError, IndexError) as e:
+			pass
 		
-		Utils.AdjustGridSize( self.grid, rowsRequired=len(model.references) )
-		for row, (reference, aliases) in enumerate(model.references):
-			self.grid.SetCellValue( row, 0, '{}, {}'.format(*reference) )
-			self.grid.SetCellValue( row, 1, '; '.join('{}, {}'.format(*a) for a in aliases) )
-			
-		self.grid.AutoSize()
-		self.GetSizer().Layout()
-	
 	def commit( self ):
-		references = []
+		self.callPageCommit( self.notebook.GetSelection() )
+		self.setTitle()
 		
-		self.grid.SaveEditControlValue()
-		for row in range(self.grid.GetNumberRows()):
-			reference = self.getName( self.grid.GetCellValue( row, 0 ) )
-			if reference:
-				aliases = [a.strip() for a in self.grid.GetCellValue(row, 1).split(';')]
-				aliases = [self.getName(a) for a in aliases if a]
-				aliases = [a for a in aliases if a]
-				references.append( (reference, aliases) )
+	def callPageCommit( self, i ):
+		try:
+			self.pages[i].commit()
+			self.setTitle()
+		except (AttributeError, IndexError) as e:
+			pass
 		
-		references.sort()
-		
-		model = SeriesModel.model
-		model.setReferences( references )
-
-#----------------------------------------------------------------------------
-
-class AliasesFrame(wx.Frame):
-	def __init__(self):
-		wx.Frame.__init__(self, None, title="Aliases Test", size=(800,600) )
-		self.panel = Aliases(self)
-		self.Show()
- 
-if __name__ == "__main__":
-	app = wx.App(False)
-	model = SeriesModel.model
-	model.setReferences( [
-		[('Bell', 'Robert'), [('Bell', 'Bobby'), ('Bell', 'Bob'), ('Bell', 'B')]],
-		[('Sitarski', 'Stephen'), [('Sitarski', 'Steve'), ('Sitarski', 'Steven')]],
-	] )
-	frame = AliasesFrame()
-	frame.panel.refresh()
-	app.MainLoop()
+	def onPageChanging( self, event ):
+		self.callPageCommit( event.GetOldSelection() )
+		self.callPageRefresh( event.GetSelection() )
+		event.Skip()	# Required to properly repaint the screen.
