+++ SeriesMgr/MainWin.py
@@ -39,9 +39,6 @@ from Results			import Results
 from TeamResults		import TeamResults
 from CategorySequence	import CategorySequence
 from Aliases			import Aliases
-from AliasesLicense		import AliasesLicense
-from AliasesMachine		import AliasesMachine
-from AliasesTeam		import AliasesTeam
 from Options			import Options
 from Errors				import Errors
 from Printing			import SeriesMgrPrintout
@@ -264,10 +261,7 @@ class MainWin( wx.Frame ):
 			[ 'points',			Points,				'Scoring Criteria' ],
 			[ 'categorySequence',CategorySequence,	'Category Options' ],
 			[ 'upgrades',		Upgrades,			'Upgrades' ],
-			[ 'aliases',		Aliases,			'Name Aliases' ],
-			[ 'licenseAliases',	AliasesLicense,		'License Aliases' ],
-			[ 'machineAliases',	AliasesMachine,		'Machine Aliases' ],
-			[ 'teamAliases',	AliasesTeam,		'Team Aliases' ],
+			[ 'aliases',		Aliases,			'Aliases' ],
 			[ 'options',		Options,			'Options' ],
 			[ 'errors',			Errors,				'Errors' ],
 		]
